requirements.txt
absl-py==1.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albumentations==0.4.6
altair==4.2.0
amazoncaptcha==0.5.7
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.5.0
appnope==0.1.4
apscheduler==3.10.4
asgiref==3.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.1
black==22.8.0
blinker==1.7.0
boto3==1.34.69
botocore==1.34.69
bs4==0.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
-e /Users/krishnanunnir/Projects/progressionmaster/venv/src/clip
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
diffusers==0.2.4
distlib==0.3.6
distro==1.9.0
django==4.0.6
django-constance==2.9.0
django-picklefield==3.1
djhtml==1.5.1
einops==0.3.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.8.0
flake8==4.0.1
flask==3.0.2
flask-cors==4.0.0
flask-s3==0.3.3
flatbuffers==2.0.7
fonttools==4.37.1
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
genanki==0.13.1
gitdb==4.0.9
gitpython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
gotrue==2.8.1
greenlet==2.0.2
grpcio==1.49.0rc3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.9.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==7.1.0
invisible-watermark==0.1.5
ipykernel==6.29.5
ipython==8.18.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.15.0
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.4
kornia==0.6.0
langchain==0.0.343
langchain-core==0.0.13
langsmith==0.0.92
lxml==4.9.1
markdown==3.4.1
markupsafe==2.1.5
marshmallow==3.21.1
matplotlib==3.6.0rc2
matplotlib-inline==0.1.7
mccabe==0.6.1
mpmath==1.2.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==2.8.6
numpy==1.26.4
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.12.0
onnxruntime==1.12.1
openai==1.45.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.0rc0
parso==0.8.4
pathspec==0.9.0
pexpect==4.9.0
pillow==9.1.1
pip==24.0
platformdirs==2.5.2
playwright==1.35.0
postgrest==0.16.11
prompt-toolkit==3.0.47
protobuf==3.19.4
psutil==6.0.0
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.3
pyarrow==9.0.0
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pycodestyle==2.8.0
pydantic==2.6.4
pydantic-core==2.16.3
pydeck==0.8.0b1
pydeprecate==0.3.1
pyee==9.0.4
pyflakes==2.4.0
pygments==2.13.0
pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytorch-lightning==1.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywavelets==1.3.0
pyyaml==6.0.1
pyzmq==26.1.0
realtime==2.0.2
regex==2022.8.17
requests==2.31.0
requests-ip-rotator==1.0.14
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
s3transfer==0.10.1
schedule==1.2.2
scikit-image==0.19.3
scipy==1.9.1
selenium==4.25.0
semver==3.0.0.dev3
setuptools==69.2.0
shapely==2.0a1
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlalchemy==2.0.28
sqlparse==0.4.2
stack-data==0.6.3
storage3==0.7.7
streamlit==1.12.2
strenum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.11.1
-e /Users/krishnanunnir/Projects/progressionmaster/venv/src/taming-transformers
tenacity==8.2.3
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2022.8.12
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0.dev20220906
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.14.0.dev20220906
tornado==6.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.19.2
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.18
urwid==2.1.2
validators==0.20.0
virtualenv==20.16.4
wcwidth==0.2.5
websocket-client==1.8.0
websockets==12.0
werkzeug==3.0.1
wheel==0.37.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
